refactor(common): move the convergence provenance vocabulary into common
The kill switch's provenance vocabulary (WORK_ORIGINS, CONVERGENCE_ORIGIN,
provenancePayloadShape, shouldHaltConvergence) lived in
apps/client/server/queueHandlers, which put it out of reach of producers outside
apps/client. packages/scripts has one such producer, and neither b4m-core nor
packages/scripts can import from apps/client - the same boundary that already put
isChunkRebuildPending in common/src/constants/chunking.ts.

Pure re-home: the module is zod-only, so the dependency cost is nil, and the old
path re-exports it so every existing importer and its tests are unchanged.

Three suites mock @bike4mind/common wholesale, which would have resolved the
re-exported symbols to undefined; they now pull them from the real module, for the
same reason the surrounding mocks already pull CONVERGENCE_PAUSED_NOTE - a literal
there would let the suite pass against a value the switch no longer recognises.
