+++ apps/client/pages/api/data-lakes/[id]/__tests__/converge.test.ts
@@ -45,7 +45,15 @@ vi.mock('@bike4mind/database', () => ({
   adminSettingsRepository: { getSettingsValue: h.getSettingsValue },
   scopedSettingsRepository: {},
 }));
-vi.mock('@bike4mind/common', () => ({ isSupportedEmbeddingModel: () => true }));
+vi.mock('@bike4mind/common', async () => ({
+  isSupportedEmbeddingModel: () => true,
+  // Real, not a literal: converge.ts stamps this on every message it enqueues and the chunk
+  // handler's kill switch reads it, so a stub here would let this suite pass against a value the
+  // switch no longer recognises. It reaches converge.ts through convergenceProvenance.ts, which
+  // re-exports the vocabulary from common.
+  CONVERGENCE_ORIGIN: (await vi.importActual<typeof import('@bike4mind/common')>('@bike4mind/common'))
+    .CONVERGENCE_ORIGIN,
+}));
 vi.mock('@bike4mind/utils', () => ({ BadRequestError: class BadRequestError extends Error {} }));
 vi.mock('@server/dataLakes/toAccessContext', () => ({ toAccessContext: h.toAccessContext }));
 vi.mock('@server/utils/sqs', () => ({ sendToQueue: h.sendToQueue }));
