+++ apps/client/server/queueHandlers/fabFileVectorize.test.ts
@@ -102,16 +102,26 @@ vi.mock('@server/utils/dataLakeSpendNotifier', () => ({ makeDataLakeSpendNotifie
 vi.mock('@bike4mind/utils', () => ({ getSettingsByNames: vi.fn() }));
 vi.mock('@server/utils/errors', () => ({ NotFoundError: class NotFoundError extends Error {} }));
 // Module-load zod schemas used by VectorizePayload.
-vi.mock('@bike4mind/common', async () => ({
-  SupportedEmbeddingModelSchema: z.string(),
-  getEmbeddingModelCost: vi.fn(() => 0.0001),
-  // Pulled from the real module rather than retyped. This handler WRITES the note and the lake-health
-  // evaluator READS it to tell a permanently-stalled file from one still indexing; production cannot
-  // drift (both import the same constant), but a literal here would let THIS suite keep passing
-  // against a string the constant no longer has.
-  CONVERGENCE_PAUSED_NOTE: (await vi.importActual<typeof import('@bike4mind/common')>('@bike4mind/common'))
-    .CONVERGENCE_PAUSED_NOTE,
-}));
+vi.mock('@bike4mind/common', async () => {
+  const actual = await vi.importActual<typeof import('@bike4mind/common')>('@bike4mind/common');
+  return {
+    SupportedEmbeddingModelSchema: z.string(),
+    getEmbeddingModelCost: vi.fn(() => 0.0001),
+    // Pulled from the real module rather than retyped. This handler WRITES the note and the lake-health
+    // evaluator READS it to tell a permanently-stalled file from one still indexing; production cannot
+    // drift (both import the same constant), but a literal here would let THIS suite keep passing
+    // against a string the constant no longer has.
+    CONVERGENCE_PAUSED_NOTE: actual.CONVERGENCE_PAUSED_NOTE,
+    // Same reason, for the provenance vocabulary convergenceProvenance.ts re-exports from common:
+    // the payload schema's fail-soft `origin` and the halt rule are exactly what the kill-switch
+    // tests below exercise, so a stub here would make them assert against themselves.
+    WORK_ORIGINS: actual.WORK_ORIGINS,
+    WorkOriginSchema: actual.WorkOriginSchema,
+    CONVERGENCE_ORIGIN: actual.CONVERGENCE_ORIGIN,
+    provenancePayloadShape: actual.provenancePayloadShape,
+    shouldHaltConvergence: actual.shouldHaltConvergence,
+  };
+});
 vi.mock('@bike4mind/fab-pipeline', () => ({
   ChunkSchema: z.object({}).passthrough(),
   EmbeddingFactory: class {
